requirements.txt
pyttsx3==2.90
SpeechRecognition==3.8.1
pipwin==0.5.1
DateTime==4.3
wikipedia==1.4.0
googlesearch-python==1.0.1